Removed utils file as per suggestion
