+++ lib/storage.js
@@ -23,7 +23,6 @@ const mime = require('mime-types');
 const validUrl = require('valid-url');
 const url = require('url');
 const Rendition = require('./rendition');
-const { fileExistsAndIsNotEmpty } = require('./utils/fileutils');
 
 const SOURCE_BASENAME = 'source';
 
@@ -95,6 +94,22 @@ function extension(filename, mimeType) {
         ext = mimeExt ? `.${mimeExt}` : '';
     }
     return ext;
+}
+
+/**
+ * Returns the time delta between two unix epoch timestamps in seconds
+ *
+ * @param {Number} path location of the file
+ * @returns {Boolean} Returns false if file does not exist or is empty
+ */
+function fileExistsAndIsNotEmpty(path) {
+    if(typeof path !== undefined && path && fs.existsSync(path)){
+        const fileStats =  fs.statSync(path);
+        if(fileStats.size !== 0){
+            return true;
+        }
+    }
+    return false;
 }
 
 // There is at least one worker (graphics magick) that in some cases depends
