+++ lib/storage/datauri.js
@@ -30,7 +30,7 @@ async function download(source, file) {
 }
 /**
  * Generate Presign Url taking local file path as input
- * Creates a unqiue name for the file, upload the file to 
+ * Creates a unique name for the file, upload the file to 
  * storage from the path provided and generates a PreSigned Url
  * @param {String} Path to local file system, where file is available
  * @returns {String} PreSigned Url of upload the source file
